MOVE interfaces from /shared to /app/shared for consistency; UPDATE component imports for files
