+++ Frontend/src/app/shared/interfaces/select-option.interface.ts
