+++ Frontend/src/app/shared/interfaces/user-interfaces.ts
@@ -49,6 +49,13 @@ export interface UserActivity {
     type: string;
 }
 
+export interface UserActivitiesResponse {
+    page: number;
+    total_pages: number;
+    total_results: number;
+    results: UserActivity[];
+}
+
 export interface UserActivityWithDaySplitt {
     date: string;
     episodeID: number | null;
@@ -66,6 +73,7 @@ export interface UserActivityWithDaySplitt {
     tracklistSeasinID: number | null;
     type: string;
     isDateSplitter: boolean;
+    picture: string | null;
 }
 
 export interface UserActivitiesPageEvent {
