+++ Frontend/src/features/user/delete-user-account-page/delete-user-account-page.component.ts
@@ -30,7 +30,6 @@ import {
     Validators,
 } from '@angular/forms';
 import { FloatLabelModule } from 'primeng/floatlabel';
-import { DeleteUserRequest } from '../../../shared/interfaces/user-interfaces';
 import { Observable } from 'rxjs';
 import { DeleteDialogComponent } from '../../sharedCOMPONENTS/delete-dialog/delete-dialog.component';
 import { VALIDATION_QUESTIONS } from '../../../app/shared/variables/validation-questions';
@@ -38,6 +37,7 @@ import { ERR_OBJECT_INVALID_AUTHENTICATION, getMessageObject } from '../../../ap
 import { ROUTES_LIST } from '../../../app/shared/variables/routes-list';
 import { UC_DeleteUserAccount } from '../../../app/core/use-cases/user/delete-user-account.use-case';
 import { UC_LogoutOfAccount } from '../../../app/core/use-cases/user/log-out-of-account.use-case';
+import { DeleteUserRequest } from '../../../app/shared/interfaces/user-interfaces';
 
 @Component({
     selector: 'app-delete-user-account-page',
